Remove ExportsHistoryMixin
The logic has been moved to the HistoryManager.
